1：rray与SString实现 2：FBX解析 3：DirectXTK不再以package存在，改为lib

+++ SandRenderer/Source/DXGIAdapter.cpp
@@ -11,7 +11,7 @@ DXGIAdapter::DXGIAdapter( Microsoft::WRL::ComPtr<IDXGIAdapter1> pAdapter )
 
 	Microsoft::WRL::ComPtr<IDXGIOutput> pOutput;
 
-	while( pAdapter->EnumOutputs( m_vOutputs.size() , pOutput.ReleaseAndGetAddressOf() ) != DXGI_ERROR_NOT_FOUND )
+	while( pAdapter->EnumOutputs( ( UINT )m_vOutputs.size() , pOutput.ReleaseAndGetAddressOf() ) != DXGI_ERROR_NOT_FOUND )
 	{
 		m_vOutputs.push_back( pOutput );
 	}
